remove support for days
As it is the goal of ths pull request to unify the format for all,
remove support for days, as periodicnoise doesn't support this unit.

Only compatibility problem is now "µs", but my Ruby foo is not enough to
support this in Ruby 1.8.7+, 1.9.x and 2.x with the same code.